add user class
